portal-python/requirements.txt
msgpack-python==0.5.6
python-dateutil==2.8.1
six==1.15.0
transit-python==0.8.302
